CSL: Allow article type to be used for preprints instead of report
Harvard (Bath) treats preprints very similarly to reports, so this is
how they have been encoded so far. The CSL specification, however, says
they should have the `article` type, so this commit adds support for
encoding them thus.

In order to ensure that the style does not break for people using
`article` in place of `article-journal`, the preprint format for
`article` only kicks in if `container-title` is missing, so the
repository name still needs to go in `archive`. Specifically, when
`article` or `article-*` are missing a `container-title`, this now
switches off rendering of subsequent "[Online]", volume, number and
pages, and switches on the printing of `status` and the publisher. With
`article-*`, a `container-title` switches off printing the `archive`;
with `article`, the `archive` is still printed and shows the "[Online]"
tag.

Closes #17.